requirements.txt
absl-py==0.7.0
astor==0.8.1
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
networkx==2.4
numpy==1.19.1
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.29.0
zipp==3.1.0
